i3: change some programs and bindings around (read body)
- Use Flameshot instead of scrot, because it has a nicer UI and actually
  freezes the screen properly.
- Change the sticky window hotkey to $mod+z because I kept hitting it on
  accident too often.
- Play the Windows XP startup and shutdown sound at their respective
  time.
- Remove tmux, because I don't really need its multiplexing. I'm using
  standalone kitty now.

+++ i3/scripts/i3exit.sh
@@ -4,21 +4,25 @@ lock() {
 	betterlockscreen -l blur &
 }
 
+play_poweroff() {
+    ffplay -v 0 -nodisp -autoexit "$HOME/Desktop/Music/shutdown.mp3"
+}
+
 case "$1" in
     lock)
         lock
         ;;
     logout)
-        i3-msg exit
+        play_poweroff && i3-msg exit
         ;;
     suspend)
         lock && systemctl suspend
         ;;
     reboot)
-        systemctl reboot
+        play_poweroff && systemctl reboot
         ;;
     shutdown)
-        systemctl poweroff
+        play_poweroff && systemctl poweroff
         ;;
     *)
         echo "Usage: $0 {lock|logout|suspend|hibernate|reboot|shutdown}"
